Addressed the issue of image size by redesigning the view page, popup problems fixed, other bugfixes
Comment section is now a fixed sidebar on the right. You can close it if you wish, it will make the image section bigger (mobile friendly).

Added a lightbox plugin for the images. You can view the image in a fullscreen lightbox by clicking the button in the top right corner of the image.

Popup content is now scrollable and has a max height. It is also wider.

Popup commenting now supports the enter key.

Fixed a bug where updating the comment section two divs with id "comment-section" were created