Update to include a third thumb
As requested by the users, a new state of review has been added to indicate whether a review needs more work. The middle thumb is now displayed on the review page and the column heading is on the review all page (no actual implementation of it yet). The database has been updated ready to store data from future implementation.