Implemented the 'Needs Work' Thumb
There is now a third thumb to register whether a user thinks a design needs more work. The values are updated in the browser based on their button selection. The boolean to work out which button is pressed has been changed to an integer, and the controller has been altered to include additional conditions.