Added threads, updated pins icons, fixed lightbox, added the ability to open pins from comment section
Threads now work. You can either add a new single comment, or reply to the previous ones. This works on the pins and in the comment section.

New custom pin icons ( Ideas? )

You now can open pins by clicking "Open pin" from the comment section. If a comment does not have a pin, the link to open the pin will not be displayed. The replies to comments which have pins also do not have the button (should it be added?).

Lightbox now works properly.