Touched up Login/Register Models
Due to combining the above pages when navigating to \Regsiter issues would start popping up. Remapped \Register to load the same cshtml content as \Login but defaulted to Register.

In the near future this will change into using _partialLayouts for the login and register forms. Within the same cshtml file.