Ability to sort comments by date added
Either select the date you wish to view in the select dropdown,or click on the date header you want to view.

Fixed the way uploaded images are displayed in the Homepage ( they used to be cropped before, now they are displayed in the correct size)

Fixed the background selector being too wide. Now it only fits the image