Fixed pin and reply issues with multiple screens.
The problem is that our usual method of refreshing the comment section automatically reloads the model, which means instead of calling the _ScreenPartial method in the controller, it calls the View method instead. This method is only meant to be used by the first screen when you load a project. I only found one way around, thats to call the screenSelector method on when replying, and that reloads all of the comment sections, including the pins, which means they get back to their initial states ( display:none for  the reply sections, all popups stay closed ) . Hopefully, these issues will be fixed after implementing Webpack into commenting.