#70 bf-caching
Fixed the issue where navigation via the back and forward buttons would cause incorrect/out-of-date content to be rendered on screen. This was an issue with Chrome and bf-cache.

Tried telling the browser to first not cache anything then not store anything via the HTTP Response headers but this did not have the desired effect and only change the behavior in FF.

Including this script in the _Layout.cshtml forces the page to refreshed if the history contains a copy. Therefore; when navigating back we get a fresh copy and when navigating forward we also get a fresh copy. No 'historic' page is ever displayed. Needed for out dynamic changes as a result of using ajax and js.

Tested on;
* Chrome 60.0.3112.90,
* FF 54.0.1