CompareTime method added
A method has been added to compare the times of the last notification sent, and the time now. If the difference is greater than 10 minutes, and the user is currently able to receive notifications it is set to true. ApplicationDbContext has also been added to the controller. When users initially login, their flag is set to true and time of notification set to when they logged in.