Removed QuickReview
Removed the option to create a different type of review. This distinction will be done server side and without user knowledge from this point on.

If no review title and description is added when creating a review, when viewing said review we assume it's a single screen and will skip the review summary page. Landing the user on the screen by default.

Whereas when a title or description for the review is provided the user will see a summary page first, and if one of the above is missing, it shows "No x provided by the author".