Working markdown in screen description.
Fixed the issue where the mardown wasn't being saved when creating a review.

The issue was that during the switch to supporting multiple screens the script to convert the markdwon to html to be saved was left behind. This also included fetching the description to save from the incorrect textarea element. Meaning plain text and not html was being stored.

Also. A new plugin has been added. 'toMarkdown' This is to enable the user to edit the description of a screen that has already been saved. This will also be used in other areas when allowing the edit of a live review, etc.

Also been added; is the markdown support for the project description. But this needs to be implemented on the front end of the view review.