Hidden comments section by default.
Nothing that big. Comments section were being expanded by default. This may have lead to some reviewers being influenced by the comments that have been made prior to the reviewer reviewing the review.

Whilst this does not address the larger issue of that a user can still expand and view the comments, this goes someway and we may address it further in the future by doing something along the lines of a BCC of a review until a user has contributred to the review.