Start of Project summary.
to-here:

Removed project info dropdown section (this will become its own summary page, therefore no need for it on the view page.

Removed reviewers section as this will feature on the review summary page also.

Change to how the comments are initally rendered when chaning to a new screen. The current state of visibilty is passed into the controller, the view html in the ajax response is changed server side depending on the state passed. Meaning if the comments are open they stay open and vice-versa when switching screens.