Staging to rebase onto markdown branch.
Changes made to the GetScreenContentFor json procedure. This now returns the screen title too. This is used to add a tooltip on hover over the screen thumbnails.

This may change in the future, for example if more information is required then maybe creating a json object may not be the best idea. Recreating the db query to return an already formed full ViewModel.