Changes to login requirements
User password can be anything now as long as it is between 8-16 characters (no need for non-letters etc..). Some changes to error messages and the way errors are displayed as well as length of username. The email address now returns an error if the email already exists in the DB.