Ability to open and close on a screen level.
Added the ability to change the screen state. If the user is the author. Adding this adds a way for the author to say that 'okay feedback has been received for this screen, I will go away and work on it' or 'cool this is all fine with this screen'.