Review author removed from the reviewers list.
When changing the reviewer selection, the author was being rendered as a reviewer in the reviewer list. To combat this simply check against the author and avoid rendering it, could in future change the server implementation to not return the author but for now, all user with access (are in the access table) are returned.