Avoid deleting a review so that it contains zero screens.
Added prevention to stop deleting the final screen of a given review, at this point the review will need to be deleted, or a new version uploaded in order to replace this screen.

We've done this to avoid the situations where a review had zero screen ids and the page was full of nulls!. If more time allowed we would spend time ensuring that the system worked with nulls but this would mean at this point in development a big chunk would need to be rewritten.