Review creation process change.
Issue: On each screen at LEAST one comment had to have been added in order for a review to created.

Cause: This was found to come from a null exception. When no comments were saved, the function to add comments to the db still ran, but on a null object.

Fix. Null check before the for each loop adding the comments. Ontop of this I've added a null check for the screen loop also.