Pinned comments can only be pinned when comments section is expanded.
Issue: found that the underling system used to pin comments more importantly what system we are using the store the comment pin location does not scale very well (honestly I don't think it does it at all, but for this more testing would be needed)

Fix: In order to get a working "version" we have had to make a design decision. That being that a new pinned comment, can only be pinned when the comments section is visible (that is the comments section is expanded) and the pinned comments are only shown on the expanded comments section view.

Why: by doing this we avoid the need for scaling, any need for storing more information in the DB. Removing this feature whilst may have a negative impact is the best worst solution for our time frame.

Learnt: a lot haha.